more cleanup

+++ cute_gemm/include/cute_gemm.h
@@ -19,6 +19,6 @@ void gemm_f32_row_row_row_row_cuda(
 
 // demo2.cu
 void gemm_f32_rrrr_cuda_v2(
-    size_t m, size_t n, size_t k,
+    unsigned m, unsigned n, unsigned k,
     const float* A, const float* B, const float* C,
     float* dst);
